requirements.txt
alpaca-trade-api>=3.0.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0
matplotlib>=3.8.0
pytest>=7.4.0
python-dotenv>=1.0.0
pytz>=2024.1
statsmodels>=0.14.0
